Re #27472 move files to widgets directory
