+++ qt/python/mantidqt/widgets/saveprojectdialog/__init__.py
@@ -5,3 +5,4 @@
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
 #  This file is part of the mantid workbench
+#
